PR revisions: FilterPanel, Stricter API, test ids, styling
